Clean up Express separation and fix server tests
- Remove old atxpServer function and tests from @atxp/server
- Extract server configuration utilities to separate serverConfig.ts
- Fix test expectations to use "Test ATXP Server" consistently
- Simplify @atxp/express-middleware to import from @atxp/server
- All tests now passing in both packages
