+++ packages/atxp-express-middleware/src/index.ts
@@ -1,10 +1,14 @@
 // Express middleware for ATXP
 export {
-  type ATXPArgs,
-  DEFAULT_CONFIG,
-  buildServerConfig,
   atxpExpressMiddleware
 } from './atxpExpressMiddleware.js';
+
+// Re-export configuration utilities from @atxp/server for convenience
+export {
+  type ATXPArgs,
+  DEFAULT_CONFIG,
+  buildServerConfig
+} from '@atxp/server';
 
 // Re-export commonly used types from @atxp/server for convenience
 export type {
